Pushed v1.27 code
- Fixed a small stuttering issue in the audio/video syncing
- Bugfix: When rendering a song to a 16-bit WAV while having audio output
  set to 32-bit, you'd never get dithering in the WAV render.
- Further optimization of the audio channel mixer (~5-16% perf. increase)

+++ src/ft2_tables.h
@@ -17,7 +17,7 @@ extern const uint16_t amigaPeriod[12 * 8];
 extern const uint16_t linearPeriods[1936];
 extern const uint16_t amigaPeriods[1936];
 
-extern const uint32_t panningTab[257];
+extern const uint16_t panningTab[257];
 
 extern const char *dec2StrTab[100];
 extern const char *dec3StrTab[256];
